修改为路径自定义高度SliceHeight、改为纵向z轴检测障碍isTraversable（实现3d避障）、添加安全阈值margin、修改空间z轴最低值ground_height、修改路径显示与路径高度一致、更新静态地图

+++ plan_env/src/edt_environment.cpp
@@ -39,7 +39,7 @@ double EDTEnvironment::distToBox(int idx, const Eigen::Vector3d& pos, const doub
 
 double EDTEnvironment::minDistToAllBox(const Eigen::Vector3d& pos, const double& time) {
   double dist = 10000000.0;
-  for (int i = 0; i < obj_prediction_->size(); i++) {
+  for (size_t i = 0; i < obj_prediction_->size(); i++) {
     double di = distToBox(i, pos, time);
     if (di < dist) dist = di;
   }
